Implement placeShip method on gameBoard class
passing an entry point and the type of ship, the method will get the
coords of the ship based on the entry point and create a new ship.
All the ships that are created are stored inside an array.
A reference to the ship is passed in the hasShip property in the gameboard
at position where ships are placed.