add placeRandomShips method on AI class
this method allow to place automatically 5 ships on the AI's gameBoard.
A forEach loop goes through each type of ship and using the placeShip
method on gameBoard class and a generated random number each ship is placed
on the gameBoard. The placeShip method may return an error when is not
possible to place a ship in the requested coord, if that happens using
recursion the function is recalled until the gameBoard includes that ship.