add sendAttack method on AI class
while the Player class receive an entry from the interface by the user
the AI needs to generate a number and then sends it to the enemy
gameBoard. The number is retrieved from an array that includes numbers
from one to ten, each time a number is generated it is spliced from
the array. This is done in order to never call a specific number
multiple times.